requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
easy-thumbnails==2.8.3
idna==3.3
Pillow==9.2.0
psycopg==3.0.16
psycopg2-binary==2.9.3
python-stdnum==1.17
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
